fix(backend): update names and scopes of getColumnFromResource and getOrgResourceTotal
Updated getColumnFromResource to getTableInfo. Use getTableInfo to fetch relevant column and table names to input as parameters for raw SQL and Prisma calls. Updated service to fetch the total[Resrouce] count per individual table; Updated getOrgResourceTotal to getTotalandMaxForResource to fetch the total count and max threshold specific to the resoureType.