fix(frontend): update Modal context for Ticket mutations
Updated Modal to apply updates using setTimeout. This was to defer the updated state to the next event loop to ensure that the modal state isn't stale; Defined record explicitly from modalProps to pass directly to TicketModal; Added useEffect to provide a live update to the Modal data upon selection to prevent stale data propagation.